Update models for ids

+++ src/main/java/com/spring/handler/datahandlerspring11/model/Bangumi.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @AllArgsConstructor
 @Builder
 public class Bangumi {
-    private int bangumiId;
+    private long bangumiId;
     private int companyId;
     private int typeId;
     private String bangumiNameJp;
