+++ src/main/java/com/spring/handler/datahandlerspring11/model/BangumiNot.java
@@ -12,8 +12,8 @@ import java.util.Date;
 @AllArgsConstructor
 @Builder
 public class BangumiNot {
-    private int bangumiNotId;
-    private int bangumiId;
+    private long bangumiNotId;
+    private long bangumiId;
     private int companyId;
     private int typeId;
     private Date bangumiAddDate;
